combine swagger and redoc docs
